Only Event Organizers Can Modify or Delete Them

+++ src/events/events.controller.ts
@@ -1,11 +1,7 @@
-import { Body, Controller, Delete, Get, HttpCode, Logger, NotFoundException, Param, ParseIntPipe, Patch, Post, Query, UseGuards, UsePipes, ValidationPipe } from "@nestjs/common";
-import { InjectRepository } from "@nestjs/typeorm";
+import { Body, Controller, Delete, ForbiddenException, Get, HttpCode, Logger, NotFoundException, Param, ParseIntPipe, Patch, Post, Query, UseGuards, UsePipes, ValidationPipe } from "@nestjs/common";
 import { AuthGuardJwt } from "src/auth/auth-guard.jwt";
 import { CurrentUser } from "src/auth/current-user.decorator";
 import { User } from "src/auth/user.entity";
-import { Like, MoreThan, Repository } from "typeorm";
-import { Attendee } from "./attendee.entity";
-import { Event } from './event.entity';
 import { EventsService } from "./events.service";
 import { CreateEventDto } from './input/create-event.dto';
 import { ListEvents } from "./input/list.events";
@@ -16,10 +12,6 @@ export class EventsController {
   private readonly logger = new Logger(EventsController.name);
 
   constructor(
-    @InjectRepository(Event)
-    private readonly repository: Repository<Event>,
-    @InjectRepository(Attendee)
-    private readonly attendeeRepository: Repository<Attendee>,
     private readonly eventsService: EventsService
   ) { }
 
@@ -38,54 +30,54 @@ export class EventsController {
     return events;
   }
 
-  @Get('/practice')
-  async practice() {
-    return await this.repository.find({
-      select: ['id', 'when'],
-      where: [{
-        id: MoreThan(3),
-        when: MoreThan(new Date('2021-02-12T13:00:00'))
-      }, {
-        description: Like('%meet%')
-      }],
-      take: 2,
-      order: {
-        id: 'DESC'
-      }
-    });
-  }
+  // @Get('/practice')
+  // async practice() {
+  //   // return await this.repository.find({
+  //   //   select: ['id', 'when'],
+  //   //   where: [{
+  //   //     id: MoreThan(3),
+  //   //     when: MoreThan(new Date('2021-02-12T13:00:00'))
+  //   //   }, {
+  //   //     description: Like('%meet%')
+  //   //   }],
+  //   //   take: 2,
+  //   //   order: {
+  //   //     id: 'DESC'
+  //   //   }
+  //   // });
+  // }
 
-  @Get('practice2')
-  async practice2() {
-    // // return await this.repository.findOne(
-    // //   1,
-    // //   { relations: ['attendees'] }
-    // // );
-    // const event = await this.repository.findOne(
-    //   1,
-    //   { relations: ['attendees'] }
-    // );
-    // // const event = new Event();
-    // // event.id = 1;
+  // @Get('practice2')
+  // async practice2() {
+  //   // // return await this.repository.findOne(
+  //   // //   1,
+  //   // //   { relations: ['attendees'] }
+  //   // // );
+  //   // const event = await this.repository.findOne(
+  //   //   1,
+  //   //   { relations: ['attendees'] }
+  //   // );
+  //   // // const event = new Event();
+  //   // // event.id = 1;
 
-    // const attendee = new Attendee();
-    // attendee.name = 'Using cascade';
-    // // attendee.event = event;
+  //   // const attendee = new Attendee();
+  //   // attendee.name = 'Using cascade';
+  //   // // attendee.event = event;
 
-    // event.attendees.push(attendee);
-    // // event.attendees = [];
+  //   // event.attendees.push(attendee);
+  //   // // event.attendees = [];
 
-    // // await this.attendeeRepository.save(attendee);
-    // await this.repository.save(event);
+  //   // // await this.attendeeRepository.save(attendee);
+  //   // await this.repository.save(event);
 
-    // return event;
+  //   // return event;
 
-    return await this.repository.createQueryBuilder('e')
-      .select(['e.id', 'e.name'])
-      .orderBy('e.id', 'ASC')
-      .take(3)
-      .getMany();
-  }
+  //   // return await this.repository.createQueryBuilder('e')
+  //   //   .select(['e.id', 'e.name'])
+  //   //   .orderBy('e.id', 'ASC')
+  //   //   .take(3)
+  //   //   .getMany();
+  // }
 
   @Get(':id')
   async findOne(@Param('id', ParseIntPipe) id: number) {
@@ -114,30 +106,46 @@ export class EventsController {
   // Create new ValidationPipe to specify validation group inside @Body
   // new ValidationPipe({ groups: ['update'] })
   @Patch(':id')
+  @UseGuards(AuthGuardJwt)
   async update(
     @Param('id') id,
-    @Body() input: UpdateEventDto
+    @Body() input: UpdateEventDto,
+    @CurrentUser() user: User
   ) {
-    const event = await this.repository.findOne(id);
+    const event = await this.eventsService.getEvent(id);
 
     if (!event) {
       throw new NotFoundException();
     }
 
-    return await this.repository.save({
-      ...event,
-      ...input,
-      when: input.when ? new Date(input.when) : event.when
-    });
+    if (event.organizerId !== user.id) {
+      throw new ForbiddenException(
+        null, `You are not authorized to change this event`
+      );
+    }
+
+    return await this.eventsService.updateEvent(event, input);
   }
 
   @Delete(':id')
+  @UseGuards(AuthGuardJwt)
   @HttpCode(204)
-  async remove(@Param('id') id) {
-    const result = await this.eventsService.deleteEvent(id);
+  async remove(
+    @Param('id') id,
+    @CurrentUser() user: User
+  ) {
+    const event = await this.eventsService.getEvent(id);
 
-    if (result?.affected !== 1) {
+    if (!event) {
       throw new NotFoundException();
     }
+
+    if (event.organizerId !== user.id) {
+      throw new ForbiddenException(
+        null, `You are not authorized to remove this event`
+      );
+    }
+
+    await this.eventsService.deleteEvent(id);
   }
 }
